Use correct model and cache key
This patch fixes the cloudformation controller by using the correct
cache key (it was using the namespace key which orders parameters in a
different way so filtering wasn't working on the controller) and also
uses the correct spinnaker model for cloud formation (it was importing
a class from the amazon library with the same name resulting in an error
deserializing the object).

Also added some logs that were useful while debugging, mostly in debug.

Signed-off-by: Xavi León <xavi.leon@schibsted.com>